Network Library > Port Scanner
Future Python Framework - Network Library

This file contains the `port_scanner` function, which is used for scanning open ports on a target IP using multiple threads.

The function takes as input the IP address to scan and the number of threads to use, and returns a list of open ports on the target IP.

The function uses the socket library to create a new socket and attempt to connect to each port in the range of 1 to 65535. 
If a connection is successful, the port is considered open and its number is added to the list of open ports. 
The function uses threading to speed up the scanning process by running multiple port scans concurrently.